+++ cartridge/argparse.lua
@@ -160,7 +160,7 @@ local box_opts = {
     replication_connect_timeout = 'number', -- **number**
     replication_connect_quorum = 'number', -- **number**
     replication_skip_conflict = 'boolean', -- **boolean**
-    replication_synchro_quorum = 'number', -- **number**
+    replication_synchro_quorum = 'string|number', -- **string|number**
     replication_synchro_timeout = 'number', -- **number**
     feedback_enabled         = 'boolean', -- **boolean**
     feedback_host            = 'string', -- **string**
@@ -419,6 +419,10 @@ end
 --      arg3: true
 --    ...
 --
+-- Each option have a type: string, boolean, number.
+-- There is an ability to set multiple types for one option.
+-- Types are split by separator ``|``,  e.g. ``string|number``.
+--
 -- @function get_opts
 -- @tparam {argname=type,...} filter
 -- @treturn {argname=value,...}
@@ -435,23 +439,38 @@ local function get_opts(opts)
         elseif type(value) == 'number' and opttype == 'string' then
             ret[optname] = tostring(value)
         elseif type(value) == 'string' then
-            local _value
-            if opttype == 'number' then
-                _value = tonumber(value)
-            elseif opttype == 'boolean' then
-                _value = toboolean(value)
-            else
-                return nil, TypeCastError:new(
-                    "can't typecast %s to %s (unsupported type)",
-                    optname, opttype
-                )
+            local multi_types = string.gsub(opttype, ' ', '');
+            local continue = true
+            local _value = nil
+            local str_value = nil
+            for _opttype in string.gmatch(multi_types, "[^|]+") do
+                if continue then
+                    if _opttype == 'string' then
+                        str_value = tostring(value)
+                    elseif _opttype == 'number' then
+                        _value = tonumber(value)
+                    elseif _opttype == 'boolean' then
+                        _value = toboolean(value)
+                    else
+                        return nil, TypeCastError:new(
+                            "can't typecast %s to %s (unsupported type)",
+                            optname, _opttype
+                        )
+                    end
+                    if _value ~= nil then
+                        continue = false
+                    end
+                end
             end
 
             if _value == nil then
-                return nil, TypeCastError:new(
-                    "can't typecast %s=%q to %s",
-                    optname, value, opttype
-                )
+                _value = str_value
+                if _value == nil then
+                    return nil, TypeCastError:new(
+                        "can't typecast %s=%q to %s",
+                        optname, value, opttype
+                    )
+                end
             end
 
             ret[optname] = _value